requirements.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.3
coverage==5.3
Django==3.1.2
djangorestframework==3.12.1
Faker==0.9.1
hypothesis==5.37.1
iniconfig==1.0.1
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mixer==6.1.3
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
pytest-django==3.10.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
wrapt==1.12.1
mysqlclient==1.4.6
django-extensions==3.0.9